nutriasistente v2/requirements.txt
openai
gspread
oauth2client
python-dotenv

